+++ packages/react-tinacms-github/src/toolbar-plugins/createGithubDeleteAction.tsx
