+++ packages/react-tinacms-github/src/toolbar-plugins/index.tsx
@@ -18,4 +18,4 @@ limitations under the License.
 
 export * from './useGithubToolbarPlugins'
 export * from './RepoInfoPlugin'
-export * from './createMarkdownDeleteAction'
+export * from './createGithubDeleteAction'
